+++ lib/middleman-asciidoc.rb
@@ -1,7 +1,7 @@
-require "middleman-core"
-require "middleman-asciidoc/version"
+require 'middleman-core'
+require 'middleman-asciidoc/version'
 
-::Middleman::Extensions.register(:asciidoc) do
-  require "middleman-asciidoc/extension"
-  ::Middleman::Asciidoc::AsciidocExtension
+Middleman::Extensions.register :asciidoc do
+  require 'middleman-asciidoc/extension'
+  Middleman::AsciiDoc::AsciiDocExtension
 end