+++ lib/middleman-asciidoc/middleman_extension.rb
@@ -1,2 +1,1 @@
 require 'middleman-syntax'
-